+++ LightCTR/train/train_embed_algo.cpp
@@ -12,7 +12,7 @@
 #include "../common/avx.h"
 #include "../util/product_quantizer.h"
 
-#define FOR_D for (int i = 0; i < emb_dimension; i++)
+#define FOR_D for (size_t i = 0; i < emb_dimension; i++)
 
 void Train_Embed_Algo::init() {
     learning_rate = 0.05f;
